requirements_installed.txt
absl-py==2.3.1
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.12.1
array_record==0.8.1
asttokens==3.0.1
astunparse==1.6.3
attrs==25.4.0
av==16.1.0
backports.strenum==1.3.1
beautifulsoup4==4.14.3
certifi==2026.1.4
cfgv==3.5.0
charset-normalizer==3.4.4
chex==0.1.85
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cuda-bindings==12.9.4
cuda-pathfinder==1.2.2
cycler==0.12.1
decorator==5.2.1
diffusers==0.35.1
distlib==0.4.0
distrax==0.1.5
distro==1.9.0
dm-tree==0.1.9
einops==0.8.1
etils==1.13.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.20.3
fire==0.7.1
flatbuffers==25.12.19
flax==0.8.4
fonttools==4.61.1
fsspec==2026.1.0
fvcore==0.1.5.post20221221
gast==0.7.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanize==4.15.0
identify==2.6.16
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
importlib_resources==6.5.2
iopath==0.1.10
ipython==8.38.0
jax==0.4.29
jax-cuda12-pjrt==0.4.29
jax-cuda12-plugin==0.4.29
jaxlib==0.4.29
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.2.4
ml_collections==1.1.0
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath==1.3.0
msgpack==1.1.2
namex==0.1.0
narwhals==2.15.0
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.15.1.9
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.28.9
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
octo @ git+https://github.com/octo-models/octo.git@241fb3514b7c40957a86d869fecb7c7fc353f540
openai==1.99.9
opencv-python==4.11.0.86
OpenEXR==3.4.4
opt_einsum==3.4.0
optax==0.1.5
optree==0.18.0
orbax-checkpoint==0.6.4
packaging==25.0
parameterized==0.9.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.1
plotly==6.5.2
portalocker==3.2.0
pre-commit==3.3.3
proglog==0.1.12
promise==2.3
prompt_toolkit==3.0.52
protobuf==5.29.5
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorchvideo==0.1.5
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scipy==1.15.3
sentry-sdk==2.49.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.2
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-graphics==2021.12.3
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.3
tensorflow-probability==0.24.0
tensorflow-text==2.19.0
tensorflow_keras==0.1
tensorstore==0.1.78
termcolor==3.3.0
tf_keras==2.19.0
timm==0.9.10
tokenizers==0.21.4
toml==0.10.2
toolz==1.1.0
torch==2.11.0.dev20260118+cu128
torchvision==0.25.0.dev20260118+cu128
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trimesh==4.11.1
triton==3.6.0+git9844da95
typeguard==2.13.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
virtualenv==20.36.1
wandb==0.24.0
wcwidth==0.2.14
Werkzeug==3.1.5
-e git+ssh://git@github.com/world-model-eval/world-model-eval.git@709665b29f5c95124e7dbc05b0d4d47cd353de97#egg=world_model_eval
wrapt==2.0.1
yacs==0.1.8
zipp==3.23.0
